requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
flask-jwt-extended==4.5.3
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3
pytest==7.4.2
pytest-flask==1.2.0